+++ packages/aimd-renderer/src/html/index.ts
@@ -9,5 +9,17 @@ export {
   parseAndExtract,
 } from '../common/processor'
 
+export {
+  createAimdRendererMessages,
+  DEFAULT_AIMD_RENDERER_LOCALE,
+  resolveAimdRendererLocale,
+} from '../locales'
+
 export type { ProcessorOptions } from '@airalogy/aimd-core/types'
-export type { RenderResult } from '../common/processor'
+export type { AimdRendererOptions, RenderResult } from '../common/processor'
+export type {
+  AimdRendererI18nOptions,
+  AimdRendererLocale,
+  AimdRendererMessages,
+  AimdRendererMessagesInput,
+} from '../locales'